THE REASON RELOADING CRASHES WAS A SIGPIPE Adding MSG_NOSIGNAL solved it Trying quicker algorithm for finding corners

+++ include/variables.h
@@ -64,7 +64,7 @@ public:
   cv::Rect        boundingRect;
   double          area;
   int             id;
-  cv::Point2d     points[4];
+  cv::Point2f     points[4];
   void NullTargets() {
     // rect = cv::RotatedRect();
     // boundingRect = cv::Rect();
